Extract "isArrayLike" to own file
