+++ lib/referee.js
@@ -1,7 +1,6 @@
 "use strict";
 
 var actualAndExpectedMessageValues = require("./actual-and-expected-message-values");
-var actualAndTypeOfMessageValues = require("./actual-and-type-of-message-values");
 var assertArgNum = require("./assert-arg-num");
 var captureException = require("./capture-exception");
 var interpolatePosArg = require("./interpolate-pos-arg");
@@ -9,7 +8,6 @@ var interpolateProperties = require("./interpolate-properties");
 var expect = require("./expect");
 var bane = require("bane");
 var includes = require("lodash.includes");
-var isArguments = require("lodash.isarguments");
 var Promise = require("es6-promise").Promise;
 
 var slice = Array.prototype.slice;
@@ -202,25 +200,7 @@ require("./assertions/is-boolean")(referee);
 require("./assertions/is-number")(referee);
 require("./assertions/is-nan")(referee);
 require("./assertions/is-array")(referee);
-
-function isArrayLike(object) {
-    return Array.isArray(object) ||
-        (!!object && typeof object.length === "number" &&
-        typeof object.splice === "function") ||
-        isArguments(object);
-}
-
-referee.isArrayLike = isArrayLike;
-
-referee.add("isArrayLike", {
-    assert: function (actual) {
-        return isArrayLike(actual);
-    },
-    assertMessage: "${customMessage}Expected ${actual} to be array like",
-    refuteMessage: "${customMessage}Expected ${actual} not to be array like",
-    expectation: "toBeArrayLike",
-    values: actualAndTypeOfMessageValues
-});
+require("./assertions/is-array-like")(referee);
 
 function exactKeys(object, keys) {
     var keyMap = {};
